ranim_items::vitem

Module geometry

Geometry items

Modules

anchor

Anchors for geometry items.

arc 🔒

circle 🔒

ellipse 🔒

elliptic_arc 🔒

line 🔒

parallelogram 🔒

polygon 🔒

Structs

Arc

An arc centered at the origin in the XY plane.

ArcBetweenPoints

An arc whose local geometry is defined by its start and end points.

Circle

A circle centered at the origin in the XY plane.

Ellipse

An ellipse centered at the origin in the XY plane.

EllipticArc

An elliptic arc centered at the origin in the XY plane.

Line

A line segment.

Parallelogram

A general affine parallelogram in local coordinates.

Polygon

A polygon with uniform stroke and fill.

Rectangle

A rectangle centered at the origin in the XY plane.

RegularPolygon

A regular polygon centered at the origin in the XY plane.

Square

A square centered at the origin in the XY plane.