ranim_items::vitem::geometry::line

Struct Line

pub struct Line {
    pub points: [DVec3; 2],
    pub extrude: [f64; 2],
    pub stroke_rgba: AlphaColor<Srgb>,
    pub stroke_width: f32,
}

A line segment.

Fields
points: [DVec3; 2]

The start and end points of the line.

extrude: [f64; 2]

The distance two endpoints extends or shrinks from its original position. Positive value means extension and negative value means shrinking.

stroke_rgba: AlphaColor<Srgb>

Stroke RGBA values.

stroke_width: f32

Stroke width.

Implementations

impl Line

pub fn new(start: DVec3, end: DVec3) -> Self

Creates a new line segment with the given start and end points.

pub fn invert(&mut self) -> &mut Self

Inverts the direction of the line segment.

pub fn points_with_extrude(&self) -> [DVec3; 2]

Returns the start and end points of the line segment considering the extrusion distance.